backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
pyflakes==2.4.0
PyJWT==2.4.0
psycopg2-binary==2.8.6
djoser==2.1.0
django-filter==21.1
django-colorfield==0.7.2
drf-extra-fields==3.4.0
drf-yasg==1.21.3
django-rest-swagger==2.2.0
gunicorn==20.0.4
python-dotenv==0.21.0
asgiref==3.3.2
requests==2.26.0
Pillow==9.1.1
pytest==6.2.4
pytest-django==4.5.2
pytest-pythonpath==0.7.4
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
isort==5.10.1